refactor: organize environmental meshes into separate files
