+++ frontend/src/components/EarthWithEnvironment/index.ts
@@ -1,0 +1,3 @@
+import EarthWithEnvironment from "./EarthWithEnvironment";
+
+export default EarthWithEnvironment;
